# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Island Dream Browser API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SearchResult {
  title: string;
  url: string;
  description: string;
  /** @nullable */
  favicon?: string | null;
  domain: string;
}

export interface SearchResults {
  results: SearchResult[];
  query: string;
  hasMore: boolean;
}

export interface PageMeta {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  favicon?: string | null;
  url: string;
}

export interface ErrorResponse {
  error: string;
}

export type SearchParams = {
/**
 * Search query
 */
q: string;
/**
 * Page number (default 1)
 */
page?: number;
};

export type GetProxyMetaParams = {
/**
 * URL to fetch metadata for
 */
url: string;
};