# lib/api-zod/src/generated/types/pageMeta.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Island Dream Browser API
 * OpenAPI spec version: 0.1.0
 */

export interface PageMeta {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  favicon?: string | null;
  url: string;
}