# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Island Dream Browser API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  title: string;
  url: string;
  description: string;
  /** @nullable */
  favicon?: string | null;
  domain: string;
}