# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Island Dream Browser API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResult } from './searchResult';

export interface SearchResults {
  results: SearchResult[];
  query: string;
  hasMore: boolean;
}